requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.2
frozenlist==1.8.0
google-auth==2.45.0
google-genai==1.56.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-google-genai==4.1.2
langchain-text-splitters==1.1.0
langsmith==0.5.2
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
orjson==3.11.5
packaging==25.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.45
tenacity==9.1.2
tiktoken==0.12.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
youtube-transcript-api==1.2.3
zstandard==0.25.0
